+++ src/router/names.ts
@@ -4,3 +4,4 @@ export const LANDING_PAGE_ROUTE = "LandingPageRoute";
 export const STORY_MODE_ROUTE = "StoryModeRoute";
 export const ORBAT_CHART_ROUTE = "OrbatChartRoute";
 export const EXPORT_SCENARIO_ROUTE = "ExportScenarioRoute";
+export const BATCH_EDIT_ROUTE = "BatchEditRoute";