+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Embed unit icons in KMZ export.
-- Basic KML/KMZ export.  
+- Basic KML/KMZ export.
 - Basic GeoJSON export.
 - Add a configurable widget/control for showing mouse position coordinates on map.
 - Zoom to multiple units (z keyboard shortcut).